Move TicketStatus updates into job
