+++ app/controllers/zendesk_webhook_controller.rb
@@ -25,59 +25,10 @@ class ZendeskWebhookController < ActionController::Base
   end
 
   def updated_ticket
-    return unless has_valid_intake_id? && intakes_for_ticket.present?
-
-    # in rare cases, there may be multiple intakes with the same ticket id
-    intakes_for_ticket.each do |intake|
-      current_status = intake.current_ticket_status
-      if current_status.nil? || ticket_status_changed?(current_status, incoming_ticket_statuses)
-        # verified_change means that we already had an old TicketStatus in the database and it had
-        # different status data.
-        verified_change = !current_status.nil?
-        new_status = intake.ticket_statuses.create(ticket_id: json_payload[:ticket_id], verified_change: verified_change, **incoming_ticket_statuses)
-      end
-
-      if new_status&.verified_change?
-        new_status.send_mixpanel_event(mixpanel_routing_data)
-      end
-    end
+    UpdateZendeskTicketStatusJob.perform_later(json_payload.permit!)
   end
 
   private
-
-  def ticket_status_changed?(current_status, intake_status: nil, return_status: nil, eip_status: nil)
-    current_status.intake_status != intake_status || current_status.return_status != return_status || current_status.eip_status != eip_status
-  end
-
-  def mixpanel_routing_data
-    {
-      path: request.path,
-      full_path: request.fullpath,
-      controller_name: self.class.name.sub("Controller", ""),
-      controller_action: "#{self.class.name}##{action_name}",
-      controller_action_name: action_name,
-    }
-  end
-
-  def intakes_for_ticket
-    @intakes_for_ticket ||= Intake.where(intake_ticket_id: json_payload[:ticket_id])
-  end
-
-  def has_valid_intake_id?
-    json_payload[:external_id].include?("intake-")
-  end
-
-  def incoming_ticket_statuses
-    # If a webhook event has an EIP status, skip reporting other statuses.
-    if json_payload[:eip_return_status].present?
-      { eip_status: json_payload[:eip_return_status] }
-    else
-      {
-        intake_status: json_payload[:digital_intake_status],
-        return_status: json_payload[:return_status],
-      }
-    end
-  end
 
   def json_payload
     params[:zendesk_webhook]
